server_api/src/modules/llm/sample/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
Cython==3.2.4
dataclasses-json==0.6.7
distro==1.9.0
easydict==1.13
ecdsa==0.19.1
faiss-cpu==1.8.0
fastapi==0.116.1
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
google-ai-generativelanguage==0.6.6
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.7.2
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
humanfriendly==10.0
idna==3.11
ImageIO==2.37.3
insightface==0.7.3
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.5.0
langchain==0.3.27
langchain-community==0.3.21
langchain-core==0.3.83
langchain-google-genai==2.0.0
langchain-openai==0.3.11
langchain-text-splitters==0.3.11
langsmith==0.3.45
lazy-loader==0.5
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==1.26.4
onnx==1.20.1
onnxruntime==1.19.2
openai==1.68.2
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
orjson==3.11.7
pandas==3.0.1
passlib==1.7.4
pillow==12.1.1
prettytable==3.17.0
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.25.8
psutil==7.2.2
psycopg==3.2.9
psycopg-binary==3.2.9
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.10.1
pydantic-settings==2.10.1
pydantic_core==2.27.1
pyparsing==3.3.2
pypdf==5.4.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.3
regex==2026.2.28
reportlab==4.2.2
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
simsimd==6.5.16
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tifffile==2026.3.3
tiktoken==0.12.0
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.3.60
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.35.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
yarl==1.23.0
zstandard==0.23.0
